102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -2,15 +2,15 @@
 
 /**
  * main - Prints the first 50 Fibonacci numbers
- * Description - a program that prints the first 50 
- * 		Fibonacci numbers, starting with 1 and 2
+ * Description - a program that prints the first 50
+ * Fibonacci numbers, starting with 1 and 2
  * Return: Always 0.
  */
 
 int main(void)
 {
 	int i;
-	unsigned long long fib1 = 1, fib2 = 2, fib_next;
+	unsigned long int fib1 = 1, fib2 = 2, fib_next;
 
 	printf("%llu, %llu", fib1, fib2);
